Bug 1191083 - Introduce pending objects support;
Pending objects, which implement GeckoThread.PendingObject, are objects
that depend on Gecko or the object itself to be in a certain state,
before its methods can be safely called. For example, objects with
native methods depend on libxul being loaded and initialized to a
certain degree; and objects that have instance native methods depend on
a C++ object being attached to the Java object.

GeckoThread.newPendingObjectBuffer produces an object that will buffer
all incoming calls on an interface that the pending object implements,
until the object's readyForCalls method returns true, at which point all
buffered calls are played back in sequence. For example, for pending
object A that depends on Gecko state S to work and implements interface
B (along with GeckoThread.PendingObject), newPendingObjectBuffer
produces a new object, which also implements B, that will buffer (i.e.
save) all calls to B's methods. Once A.readyForCalls(S) returns true,
all previously buffered calls will be played back on the original object
A, in the sequence of the original calls.