Bug 1391268 - 2. Expose mozglue linker error message to Java code; r?glandium
This patch makes us save the last message from an ERROR() statement, and
use that message when throwing an exception to Java code, in order to
make lib loading crash messages a lot more useful. This does make
ERROR() not thread-safe, but I think that is okay since it's only used
by the single-threaded linker.